+++ substrate/frame/examples/mock-builder/Cargo.toml
@@ -15,31 +15,31 @@ targets = ["x86_64-unknown-linux-gnu"]
 [dependencies]
 codec = { package = "parity-scale-codec", version = "3.6.1", default-features = false }
 scale-info = { version = "2.10.0", default-features = false, features = ["derive"] }
+
 frame-benchmarking = { path = "../../benchmarking", default-features = false, optional = true}
 frame-support = { path = "../../support", default-features = false}
 frame-system = { path = "../../system", default-features = false}
-sp-io = { path = "../../../primitives/io", default-features = false}
-sp-runtime = { path = "../../../primitives/runtime", default-features = false}
-sp-std = { path = "../../../primitives/std", default-features = false}
+
 polkadot-runtime-common = { path = "../../../../polkadot/runtime/common", default-features = false}
 
 [dev-dependencies]
-sp-core = { path = "../../../primitives/core", default-features = false}
-mock-builder = { git = "https://github.com/foss3/runtime-pallet-library", branch = "polkadot-v1.2.0" }
+sp-io = { path = "../../../primitives/io"}
+sp-core = { path = "../../../primitives/core"}
+sp-runtime = { path = "../../../primitives/runtime"}
+sp-std = { path = "../../../primitives/std"}
+
 polkadot-primitives = { path = "../../../../polkadot/primitives", default-features = false}
+
+mock-builder = { git = "https://github.com/foss3/runtime-pallet-library", branch = "polkadot-v1.2.0" }
 
 [features]
 default = [ "std" ]
 std = [
 	"codec/std",
+	"scale-info/std",
 	"frame-benchmarking?/std",
 	"frame-support/std",
 	"frame-system/std",
-	"scale-info/std",
-	"sp-core/std",
-	"sp-io/std",
-	"sp-runtime/std",
-	"sp-std/std",
     "polkadot-runtime-common/std",
 ]
 runtime-benchmarks = [
